Check if any registrations include the client_id
Based on the 5 minute tutorial, you are adding a second Application and authorizing the user to that Application. The existing python demo only checks if the first registration matches the client_id, and doesn't take in to account if other applications may be registered. To make this code work out of the box for the 5 minute tutorial, this change will check if any of the registrations match the client_id, as opposed to only the first registration returned.